Add tests for runtime types
